requirement.txt
Flask
Flask-Login
Flask-Mail
Flask-SQLAlchemy
flask-restful
redis
# elasticsearch==7.12.1
pycryptodome
pydes
aes



